+++ client/UI/UITypes.h
@@ -30,4 +30,31 @@ namespace UI
         TOOLTIP,
         MAX
     };
+
+    // Text
+    enum class TextHorizontalAlignment : u8
+    {
+        LEFT,
+        CENTER,
+        RIGHT
+    };
+
+    enum class TextVerticalAlignment : u8
+    {
+        TOP,
+        CENTER,
+        BOTTOM
+    };
+
+    struct TextStylesheet
+    {
+        Color color = Color(1, 1, 1, 1);
+        Color outlineColor = Color(0, 0, 0, 0);
+        f32 outlineWidth = 0.f;
+
+        std::string fontPath = "";
+        f32 fontSize = 0;
+
+        f32 lineHeightMultiplier = 1.15f;
+    };
 }
